Fixed a bunch of warnings under Windows, some of which even made sense.
git-svn-id: http://svn.berlios.de/svnroot/repos/oolite-linux/trunk@1294 127b21dd-08f5-0310-b4b7-95ae10353056

+++ src/Core/Entities/OOEntityWithDrawable.m
@@ -59,7 +59,7 @@ MA 02110-1301, USA.
 }
 
 
-- (GLfloat)findCollisionRadius
+- (double)findCollisionRadius
 {
 	return [drawable collisionRadius];
 }
